Configure position independent code
It may not be needed everywhere as it is configured now, but it's a
hammer to the "testing on Linux" problem. In the best scenario, Gradle
should generate both variant automatically unless explicitely defined.
Then, whatever is required would be resolved automatically through
attribute matching. There are some compatibility rules that can be used
to match between PIC and non-PIC code in specific scenario. This domain
knowledge should be coded within Gradle to avoid confusing the users
with this decision.