Seperate the asciinema execution into different home directory
There seems to be a concurrent execution issue in Asciinema that prevent
multiple asciinema process to run at the same time on a clean system. It
cause a file to already exsits which is most likely being created by
another asciinema process which another one already detected the
file/directory is missing. It seems to be related to the "install ID"
for asciinema. Since we aren't submitting the recording but keeping the
.cast file locally, we don't really care about this "install ID". Let's
try to fake a-user-per-asciinema-instance.