Allow users to configure compiler and linker arguments
There are some limitation to the approach caused by the hoop that needs
to be jumped through in order to minimize API leakage. As Nokee break
free from the Software Model, a lot of flexibility will be gain. For
now, it should be good enough for the majority of the use cases.