Migrate to the new toolbox conveniences
There is no need to declare groovy-all and kotlin-libstd dependencies
when using the Gradle API outside the Gradle environment (the build tool
itself) as the gradle-api artifact properly declare the compile time and
runtime dependencies.

Functional test dependencies can now be declared on the functionalTest
component automatically created by *-gradle-plugin.

You can now declare TestKit-ish runtime dependencies into the
pluginUnderTestMetadata via the dependency block of the functionalTest
component. I say "ish" here because this pluginUnderTestMetadata should
probably be on the gradlePlugin's dependency block which doesn't yet
exists. This means that adding multiple
GradlePluginSpockFrameworkTestSuite would share the same dependency
which is somewhat what we want but would also remove the possibility of
testing a plugin against various version of a plugin. This is a core
issue to the `java-gradle-plugin` which the toolbox won't solve at the
moment. Truth to be told, Nokee doesn't have that requirement.