Ensure support for JNI library implemented in Kotlin
The commit includes fixes to the JNI plugin to behave properly with the
fact that the Kotlin plugin creates an `api` configurations but also
resolve the `jar` task early.

It also contains a sample demonstrating Kotlin/C++ JNI library. A fix to
the Jar ANTLR output parser had to be done to support nested class file
name (containing $). Finally, the test code was shuffle around to allow
some testing for Kotlin language.