Implement a first draft of the core execution APIs
The core execution APIs aim at solving several use case around spawning
and handling processes. The iOS support needs to delegate to a lot of
tools that needs to be found as well as executed. The execution varies
in complexity between using the "simple" process builder to fanning out
using worker action. The documentation process is also heavy on spawning
processes, sepecially the asciinema generation. The sample tests mostly
requires a way to compare similar capability tool's output to assert the
wellbehaving of the sample steps. Executing Gradle within fixture also
has similar requirements in the toolbox. Finally, the compiler
infrastructure will be breaking its chain from Gradle core which again
needs a better modeling for invoking the compilers. It's also beneficial
to ensure we can query the entire command line for features like
generating the compile database.

This API aims to be public but shouldn't be used until I work with it
for a little longer en ensure the bugs are iron out.