Improve the embedded server by splitting up framework and tool locator
This commit introduce the concept of runtime. We have 3 as of now: base,
native and darwin. The base runtime provide the support for the embedded
server. The native runtime provides strictly native types that defines
native component at runtime. Finally, the darwin runtime provides the
extensions to the base runtime specifically for the macOS, iOS, etc
runtime. At the moment, it provide the framework and some iOS
tool locators.