Add all variant instead of trying to minimize which one we create
With the recent change, the variants are resolved only as needed so
there is some sort of optimization in avoiding unbuildable variant.
However, the direction we are taking is to allow users to configure
additional toolchain, say GCC on docker, to compile an unbuildable
variant. On top of that, it's better to give access to all variants to
the users instead of trying to limit them. History has shown that users
will start to depends on the internal implementation of the plugins
which we want to avoid. All the variants are available, but we force
them to be lazy. We could for them harder but it's alright for now.