Break the chain from the software model configuration
By breaking off from the software model, we had to introduce some
convenience types, mostly the SourceSet and SourceSetTransform concept.
They were successfully test drove in the documentation plugin and seems
to help improve the situation. Note those new conveniences are not
public yet but will become as we move toward a composable interface. All
of the JNI library plugin is off the software model with the forever
exception of the toolchain. With the previous work with improving the
Nokee embedded server, we will soon migrate all the toolchain discovery
in there and shim the ToolProvider with Gradle. The compiler
infrastructure will stay for a bit longer.