Formalize the dimension concept
There is still some work left to be done, however, this commit show the
direction Nokee is going for the manipulation of dimensions. Each of
them will be typed to allow a strong model. The variant aware component
will, at their core, deal with build variants, a raw/high level
reprensentation of the variant to create. For high-level component, it
will be possible to manipulate the build variants directly. Each
dimension declared needs to be present once. Two build variant can't be
the same (collision). Each dimension value can be subtype of the
dimension type.

Lower-level component can impose restriction on the direct modification
of the build variants as well as provide convenience. For example, the
target machines idea from the new core native plugins, but we could also
provide build type as another dimension.

Although I don't personnally like the idea of variant filtering from
Android, I may offer it with a warning. It becomes hard to deal with
scenario like: I want debug, build optimized and release build type
except for that specific variant that I only want release build type.
The convenience will be introduced as need when we are going to use the
plugins.