Publish an all inclusive artifact to help with aligning the dependencies
In general, the artifact id are an implementation detail. Users should
use `nokee-gradle-plugins` artifact which will include all the other
artifacts. If for some reason the users wants to depend on individual
artifacts (not recommended and won't be documented), they can use
`nokee-platform` as a BOM/platform to align the dependencies.