Add support for macOS SDK version different than current OS
It is especially useful when using older version of Xcode. This commit
narrow the use case to only macOS SDK. Before we can expose all SDKs, we
need to add attributes to the artifacts dicovered so they can be
differentiated. For example, version 9.3 for iOS or tvOS or macOS is not
the same.