Improve error message from core execution
It's a limited change for quick convenience. Core exection will evolve
into a much better convenience accounting for more use cases. This
change capture all of the process output and append it to the exception.
A better alternative would be to capture a rolling buffer of the last X
lines and display that instead. It needs to account when user wants to
discard the output, capture it, redirect to stdout, etc.