Formalize the component layer
The component layer is an important distinction between the extension
and the actual component. In theory, a variant could be a component as
well (one dimensional component). It's basically the difference between
an variant aware vs single variant component. The concept is kept
internal at the moment as there are still some open questions to
answer.