Add support for source layout configuration on all native plugins
JNI and iOS plugins are excluded at the moment. There isn't a good
reason for excluding iOS plugins aside from the need to have more tests.
For JNI, it's a different story, we need to expose the source set
container which doesn't exists, yet. All of the source set needs to
become public API which means more thoughts needs to be put into it.