Ungradlefy most of the codebase
The reason for ungradlefying the codebase is flexibility. Gradle
provides some conveniences to the plugin authors but those are either
limited, imcomplete, or too rigid for our need. We are left to manually
implements everything as the "Gradle-way" is just different enough to be
incompatible with other framework such as Dagger which we will be
adopting. Going forward, we are going to engineer Nokee using common
software engineering practice instead of using the opinionated Gradle
practice. All this work will be in sync with Gradle idiomatic way of
developing with Gradle meaning it should not affect performance,
usability and DSL compatibility.