Add factories to decorate the extensions
Although this commit adds quite a bit of code, it serves as an
exploration to eliminate the issue with ObjectFactory. The
ObjectFactory cause runtime exception instead of compile-time exception
when the instantiation point and constructor argument list becomes
out-of-sync. Two solution are now being explored, 1) generate factories
with custom instantiator (aka the ObjectFactory) and 2) compile-time
decoration of the types instead of relying on the runtime decoration
done by the ObjectFactory.