Rework component dependencies into the three bucket type
This refactor brings the "automatic namespacing" concept closer to the
finish API, it still have some wierdness that needs to be clean up.
Basically, any entity collection should have an idea of who owns them.
Anything created underneath knows about the owner and can generate flat
names when needed (e.g. configuration and tasks). It will also allow us
to group all entities in a master collection that we can then query
according to the owner making the distinction between all binaries of a
component vs the binaries of a specific variant.