Ungradlefy the Groovy DSL mixin of public types
As we are moving away from dynamic instantiation, we need to manually
mixin the Groovy DSL methods inside the public types. It can be
perceived as unecessary or going against idiomatic Gradle but is in fact
much better this way as IDEs will have better auto-complete as well as
eliminating a whole class of errors because of the dynamic
instantiation.