Use the new lazy Nokee types in the store and test suite container
We got an opportunity to eliminate the custom implementation of
DomainObjectCollection and related types. The store and test suite
container are now managable independent implementation thanks to the
Nokee types and utilities. The store should evolve into the
ComponentContainer and some sort of global internal container for all
IDE-able components. Those are two concerns and should be decoupled in
later commits.