Introduce task registry
The task registry is based on a hackathon at Gradle to link several
container together. It focus solely on registering tasks. It mix in
additional concept such as tracking the owner of the task via the
identifier. Later, we will publish events from the registry to notify
anyone of interest when a task is discovered and registered. It also
gracefully handle "register if absent" the lazy equivalent of
maybeCreate, a convoluted missing feature with the lazy task work.