Remove the custom code for attaching build scan to commit
A recent change to GitHub Action jump straight to the end of the build
log as well as rendering build scan as clickable links. This superseed
the fundamental issue we tried to address with this custom code. Also,
the code never worked properly as the GitHub Action for Gradle often
missed the build scan URL and didn't extract it properly.