Allow source set adapter to configure using relative String path
This feature is to support the JVM source set adapters. These source set
are a bit special as they are trying to adapt the Gradle source set
concept into the Nokee source set. It sort of square peg into round
hole. It's not properly working and in sync. The tests were disabled and
serve only a purpose of exposing the files. We may want to move toward
an stict read-only adapter instead of a read-write adapter.