Increase backoff/retry configuration
On GitHub CI, we are seeing quite a few connection reset. There is quite
a few reason why those may be happening. The current feeling is JCenter
is a bit flaky from time to time and we are failing builds because of
it. A better solution would be to have a on-site cache but GitHub Action
don't allow them so we are going to try some backoff and retry
configuration.