Introduce doSomething* model action factory methods for tests
Instead of using doNothing() model action factory method, we use a "do
something" model action concept. In the end both does nothing but for
the tests it better convey the intention. The reason being that we may
treat "obvious" do nothing action specially by ignoring the entire
action registration as there are valid cases that arise. To keep the
flow going, we may simply return a do nothing action from functional
methods that would then get ignored as an optimization later on.