Fudge the created state to make way for the initializer action
The initializer actions can add more projection to a node. Once the node
becomes initialized, the projections are final, no more change, done.
It will allow for NodeInitializer to exists which will be a sort-of
registerFactory/registerBinding API for the entire model. It use normal
actions. The advantage to reuse model action is the ability to create
ultra specific node easily (for testing) while reusing the same building
blocks to register/construct node based on public type alone.