Introduce registry for project configuration
Instead of mixing dependency bucket and Gradle configuration concept, we
will split them into two different concept. The project configuration is
a shim on top of Gradle configuration allowing to create configuration
if absent as well as handy helper for configuring Gradle configuration
upon creation or assert the configuration match what is expected.