Introduce the first draft of documentation kit
The documentation kit will take care of everything a Gradle plugin needs
in term of documentation. This draft contains support for DSL and API
reference as well as a Javadoc rendering for API reference. It also use
the Nokee universal as a trial run to see the rough edges. These
documentation components are also distributable across multiple
projects. DSL and API are now handled by each individual projects. It
will soon be the same for all samples and user manual.