Introduce a GitHub Pages site plugin
This plugin is one step further then where we want to be. There should
be a site-base/site plugin that take cares of the lifecycle task. The
GitHub Pages site plugin simple handle the publishing and assembling of
the site. Rendering of the site should be handle by another plugin.