Pass the second parameter as first parameter for single argument closure
The closure method is a Groovy convenience for
ModelNode#node(Object, Type BiConsumer). As the delegate is already
configured as the ModelNode instance, we can "push" the second argument
in a two argument closure to the first argument in a single argument
closure. This is done strictly for convenience given access to the model
node via delegate and known object via 'it' argument.