Move all dimension/build variant handling code to coordinates
It offers a more robust implementation. More features will be included
as we move forward. The goal here is to create a flexible concept to
allow custom dimensions to be declared and included in a component. It
would also be possible to use the coordinate code to handle the heavy
lifting for managing variants.