Remove assertable action
For now, we will back peddle on this concept. It would be much better if
we can assert a domain object is properly configured before reusing it
but it become quite confusing quite fast so we will need to revisit this
concept at a later time. Regardless of what we do, we will need a richer
type that is available via model-core and core-utils.