Reimplements Gradle toolchain plugins to avoid calling xcrun on Windows
When probing macOS toolchain, Gradle will invoke xcrun to query the SDK
path regardless if the system has or not the tool installed. It result
in a execution failure bubbling up all the way to the project and
failing the build. We could catch the exception and silence it but we
can get the supported platform from the toolchain implementation which
we exposed in Nokee's custom toolchain implementation. We will move to a
better implementation at a later date.