Update Gradle development samples
There are several things at play here. As part of the latest toolbox
upgrade to support Gradle 7+, we updated the spock version of the
fixtures. The upgrade force us to update Spock in our samples. However,
newer Spock BOM have Gradle module metadata causing a conflict on the
minimum supported Java version. Rightfully so, the builds now fails when
using minimum Java version for Gradle 4.9 (which is Java 1.7). For this
reason, we use Gradle 5.1 which has a minimum Java version of 1.8. Some
outputs needs to be updated as well.