Mark all ModelNode DSL related type as deprecated
The ModelObject concept replace the ModelNode DSL concept. The reason
being that ModelNode DSL, although useful on paper, always requires the
KnownDomainObject instance making the configuration more complicated.
Instead, the ModelObject represent an unrealized domain object that can
be configured at an higher level (the object layout and relation with
other object). In fact, the ModelObject is loosly a schema-less mapping
to the underlying Graph. As it stands now, this claim is not 100%
accurate but it's the direction we will move forward.