Introduce smoke test for the Nokee coordinate from the plugin portal
There was an issue introduced in the plugin portal for plugins
previously hosted on Bintray. The smoke test verify the dependency
insight as seen from Gradle.

There also a draft for build script modeling to support dual DSL
generation as well as a more fluent script declaration when using
runnerkit. The draft is based on the polyglot test fixtures from Gradle.