Details a C application with distinct property and source set entities
Mixing both type of entities is wrong in our opinion. The source set
entities belong to the graph of domain objects. The properties belong to
the graph of element. Both type of entities can have projections which
should be lazily realized as needed. The same can be true regarding
element and domain object discovery.