Move variant calculation into the model registration
Despite the fact the code looks worst than before, this commit is an
intermediate steps to migrating variant calculation into the model.
There are several feature gap that needs to be addressed as well as some
consolidation to do in term of business logic. The goal is to eliminate
the current NativeApplicationComponentVariants implementation in favor
of a more concise value object holding the variants. We also want to
eliminate the double component object as well as the extension
finalizing methods.