Unpack delegted projection
When searching for a projection, the type we use may not be the correct
type. For this reason, we need to use the type known to the projection.
For delegated projection, the type is simply the type of the first
projection wich is mostly right but not always right. This commit will
create fake projection that will bridge the gap. We should really just
add the fake projection directly to the model property but that is quite
complicated to get right at the moment as it requires quite a bit of
gymnastic. We should use a model stream (taken from the other master PR)
for an entity components (instead of a Java stream) so we can simply
bridge both entity in terms of components. In the event we support
removing components, we may have a small issue with using the model
stream as it doesn't support the "removal" of elements. We would have to
wrap the add/removal into an event and stream those instead.