Try to use NamedDomainObjectProvider if projection present
We will most likely allow a provider projection automatically for any
domain object. Until we get this implemented, we will search for the
right provider manually. We should also forward as much as possible any
mutation action directly to the provider instead of through the model as
it will generate less single use action.