Introduce a more rich language source set experience
The code here is not tested and serves only as the base implementation.
Future commit will include test coverage to ensure everything works as
expected. In a nutshell, the language source set will now include the
headers/modules, configuration and compile task. The source set should
also be register on the variant instead of the component.