Improve component reference for projection
A projection can be "view" as different Java object. For most domain
objects, we can view a projection as Provider, configurable Provider
(NamedDomainObjectProvider), KnownDomainObject and actual object. Each
one are valid. This commit doesn't account that some projection are
realized when the domain provider is realized. For this reason, we use
the honor system to only "view" a projection as the actual object when
the state is realized or higher. Later we will disallow this or maybe
provide a "early" realize which have some constraint on the valid action
on the object.