Further simplify language source set rules
Instead of going for a component-based approach during domain object
registration, we choose to use a projection-based approach. Both
approach are fine, however, given we need to support vanilla Gradle
appraoch, the projections are the driving force. For this reason, we now
use MixIn interface to detect which rules to apply on which entity. It's
a bit more streamlined with how Gradle works at its core. More changes
are coming to simplify the rules registration.