Return allow KnownDomainObject to be convertible into configurable provider
It only make sense to allow converting to NamedDomainObjectProvider.
Note that we don't forward everything to the provider as we believe the
provider and the configuration are two distinct operation on the known
object. We can always implement a ConfigurableStrategy to use the
ConfigurableProviderConvertibleStrategy which is an implementation
detail for KnownDomainObject implementation.