Use FullyQualifiedNameComponent
Despite the difference between FullyQualifiedNameComponent and
FullyQualifiedName is very slim, we still want to start introducing a
difference between a component in the ECS sense versus a value object
representing a domain model. We believe that we may later want to
optimized how components in our ECS are stored and managed which would
further confirm our distinction choice.