Introduce child element query on model element and object
Instead of having to executed complex query against the model registry,
these new APIs introduce a convenient way to access child elements. It's
similar to the links from the software model. The API differs from the
property query in the sense that it allow any child element to be
queried where as property query will only return model property
instance. It's expected to throw an exception if multiple elements with
a matching projection are found.