Expose component bits on component reference
There is a distinction that isn't clear in the current implementation
between the component type bits and the component family bits. The first
one is a single bit that represent the exact type while the second one
is series of bits that represent the full hiearchy of the type. It
includes the super types and interfaces. In theory, there shouldn't be
much hiearchy in the component types. However, we made a decision early
on to allow any Object as components which means we can literally have
anything as component. We are working on making a clear contract for
components which will clean up the component bits a lot. For example, we
accept Gradle custom types which include a bunch of internal interfaces.

Signed-off-by: Daniel Lacasse <daniel@nokee.dev>