Introduce matcher for no value Gradle provider/property
Asserting that a Gradle provider or property has no value is a bit
verbose as it rely on checking exception message. The exception message
is also different between property/provider type as well as if they have
a custom display name. To make the matter worst, they also differ if the
provider has a mapping chain. The match takes care of all this for the
user.

Signed-off-by: Daniel Lacasse <daniel@nokee.dev>