Introduce ModelComponent base interface
The interface will allow for more optimization and structure around
entity components. Initially, we thought on allowing any Object as
components. However, there is a clear distinction between what is a
component and everything else in our domain design. There is also
opportunities to treat the component differently allowing for better
performance.

Signed-off-by: Daniel Lacasse <daniel@nokee.dev>