Introduce model action as entities
There are some important distinction to make which this commit may
introduce confusion. The ModelAction prior to this commit should be
consider as System in the Entity-Component-System architecture. The
ModelAction in this commit are the bridge between Gradle's Action and
the ECS. The ModelSpec prior to this commit will be phase out and
removed. As mentioned, the old ModelAction will evolved into fully
fledged System which will strictly act on components. The new
ModelAction will act on projections. Starting now, projections should
only be a glob of components. Most of the implementation of the new
ModelAction is private on purpose. Users should use factory methods to
create the action registration. The new ModelSpec only serves as a smart
filter. It is not a complete replacement on the vanilla Gradle Spec
type. In fact, the ModelSpec act at the component level, while the Spec
act at the projection level.

Signed-off-by: Daniel Lacasse <daniel@nokee.dev>