Fold multi-value configurable identity into the component value itself
The identity holding all values is not aware when multiple values are
used. This avoid using a custom type to hold the values. The goal around
reconfiguring the internal specification implementation is to allow
some ad-hoc values to be included to allow cross-cutting configuration
such as the compileTasks on binaries.

Signed-off-by: Daniel Lacasse <daniel@nokee.dev>