Force realize component elements properties when owner realized
With the new component elements properties domain entities are no longer
nested under the property themselves. Instead, the property is
disconnected from from the entities making it important to force realize
the property to kick off any configuration action that may be queued
underneath. It's only a temporary solution until we can fully migrate
everything. We may just execute the actions on the view right away as
the views should almost be treated like Gradle Property.

Signed-off-by: Daniel Lacasse <daniel@nokee.dev>