Introduce a more composable Gradle build script model for both DSLs
Being able to model Gradle build script for both DSLs is very important
for various piece of Nokee's infrastructure. First, we want to move
away from Spock framework which will require some ways to model build
script in pure Java that is convenient. Then we have various build
script snippet modeling that we need to account for such as the plugin
management block for the samples. Finally, we want to introduce a
buildInit equivalent for Nokee plugins which will require some sort of
build script modeling.

The modeling introduced in this commit is a bridge between Gradle's
buildInit modeling and dependency management functional testing
fixtures. It focuses on three concepts, the expressions, statements and
blocks. The expressions allow a mapping between different syntax
(currently only Groovy and Kotlin). The statements are the various
section a build script may contains. Finally, the blocks are Gradle
specific build script segment. The blocks have some opinion on which
other block can be mixed together. They can also stand on their own.

Signed-off-by: Daniel Lacasse <daniel@nokee.dev>