Allow prefixing statement with variable
This is useful for the odds Gradle APIs that does not have conventional
delegating strategies. In those cases, either we explicitely prefix the
statements with `it.` or we name the first parameter and use that new
variable to prefix the statements. The `settingEvaluated` and
`beforeSettings` on `Gradle` instance are example of such odd APIs.

Signed-off-by: Daniel Lacasse <daniel@nokee.dev>