Rewrite the project.pbxproj writer using an intermediate model
The pbxproj file is not a straight forward PBXProject serialization.
There is an intermediate transformation which convert the PBXProject
into a simpler dict/array/primitive/reference bag of objects. This
transformation allocate Global IDs which are used inside Xcode schemes.
Given we want the model to be immutable once created, we can't simply
mutate model to add the Global IDs like we used to do (aka what Buck
do). Instead, we flush down the model into a more primitive one which
includes the Global IDs and references to other objects.

Signed-off-by: Daniel Lacasse <daniel@nokee.dev>