Rewrite ASCII plist grammar
The new grammar is more strict on what it matches. Although we could
match integer, real and boolean data type, the information about the
ASCII plist standard is says those data type are interpreted as String
in ASCII format and the burden relies on the application to properly
cast the data type.

Signed-off-by: Daniel Lacasse <daniel@nokee.dev>