Allow root project to be injected as parameter in JUnit 5
We also keep the same Project instance through out a test execution
meaning if the Project instance you get from field injection and method
parameters are the same. We will, at some point, allow to parameterize
the injection so you can build a graph of project but for now only a
single root project is used.

Signed-off-by: Daniel Lacasse <daniel@nokee.dev>