Move language source set creation based on tag
We tried to solve this issue multiple time which is not so trivial. We
want some way to "automatically" create source sets while avoiding
overstepping boundaries causing source sets to be created in context
where it doesn't make sense. For example, both a CApplication and
CppApplication was present in a project, we should not see a C++ source
set in the C application and vice versa. The idea behind a tag is to
contextualize the automation. If we apply it to a project entity, then
everyone will have the source set but if we apply it to a component then
only the entities owned by that component will have the source set. It
should provide a good balance.

Signed-off-by: Daniel Lacasse <daniel@nokee.dev>