Allow for functional ifPresent on Provider
The new utility is more performant then the hand rolled if isPresent()
followed by value query get() because both isPresent() and get() will
calculate the value which would normally result in the provider's
Callable to be called. Instead, ifPresent uses getOrNull() which we then
check if the value is null for presence removing one value calculation.

Signed-off-by: Daniel Lacasse <daniel@nokee.dev>