Create execution plan for the rules before executing any rules
We need to check the rules against the _current_ entity state to narrow
down the "rules to execute" before any rules are actually executed. As
the universal model is highly reactive, we could capture a rule that
isn't executable now but becomes executable when a previous rule
executes. The issue is the "becomes executable" will be handled as part
of another execution plan (a new execution plan is created on each
entity change). The rule will then execute twice, once in the newer
execution plan and a second time "while we are executing overlapping
rules" in an early execution plan. Instead, in this commit, we precheck
the execution plan to keep only the rules that matches before we start
reacting to everything. We check a second time the compliance of the
rule before we execute them. This may be wrong but for now we will do
that and revise later.

Signed-off-by: Daniel Lacasse <daniel@nokee.dev>