Remove restriction to PBXObject in coders
The idea is to keep the coders as one layer of indirection aka immediate
object encoding/decoding. Some object are encoded inline while others
are encoded as separated objects. In theory, it shouldn't matter from
the encoding/decoding point of view which is which. Thus, removing the
restriction seems like a good idea.

Signed-off-by: Daniel Lacasse <daniel@nokee.dev>