First pass into implementing adhoc repositories
An adhoc repository behave simillarly to a Maven repository except the
hosted data is dynamically generated. For example, we can create a
repository for the various toolchains or specific libraries like Windows
SDK/DDK. The goal is not to allow users to declare their own adhoc
repository but instead build infrastructure on top of Gradle dependency
engine to support the various third-party repositories.

Signed-off-by: Daniel Lacasse <daniel@nokee.dev>