Introduce an universal-model plugin for Nokee build
The new plugin try out a different approach to our own Nokee build
plugins. Instead of relying on _broken_ Gradle core plugins or on a
hiearchy of base/conventional plugins, we rely on composable actions
that abstract away most of the implementation details. Then, it's a
matter of choosing the right combination of actions to execute. It would
allow for greater flexibility and cherry-pick what each plugins does
instead of fighting with opinionated rules.

Signed-off-by: Daniel Lacasse <daniel@nokee.dev>