Extract variant capability from platform-base into independent module
The goal for those independent modules is to contain the responsability
of each capability. In simple terms, a universal model capability is a
set of rules that comfort to specific components of an entity. Then each
entity add the required components to get the expected result. This
specific capability act on the variants. Here variants can mean build
variants, test variants or any other variant aware in the general sense.

Signed-off-by: Daniel Lacasse <daniel@nokee.dev>