Introduce a keyed object layer to support reencoding
The main issue with reencoding is finding out about the known references
which may or may not still exists depending on the mutation of the
PBX/XC models. We keep track of the global IDs so if we change an
element of a known global ID, we expect it to be modified for everyone
referencing that ID. For now, the layer only forward to the delegate.

Signed-off-by: Daniel Lacasse <daniel@nokee.dev>