Squash coding keys to encode so only the latest one is encoded
We may have several layers of keyed object (when mutating them for
example) and we only want to encode the latest value for each keys. For
this reason, we will accumulate them and flush the encoding once we are
mostly sure that we have the full representation of the object.

Signed-off-by: Daniel Lacasse <daniel@nokee.dev>