Allow PBXProject to be converted back to an builder
This feature will be implemented by all Xcode model, PBXProject is just
the first one. The idea is to allow immutable manipulation of a model
(create a derived model without changing the original one). We also want
to keep any keys that may not be understood by our models. Finally, we
want to keep the ordering the same. For this reason, we create layers of
keyed object where the latest value is the one we use and present to the
user. We still have the issue where other model maybe referencing older
model which will keep returning these older model. That is an issue at
the unarchiving level. We should allow some sort of layering for the
backend as well.

Signed-off-by: Daniel Lacasse <daniel@nokee.dev>