Allow full reencoding of Xcode model
The reencoding has some minor issue around swapping the implementation
of an model for a newer one. Not all reference are updated.

The commit also include an experimental Hamcrest matchers for
invocation result. An adapter exists for the Mockito allowing to verify
invocation using Hamcrest making the assertion more uniform.

Signed-off-by: Daniel Lacasse <daniel@nokee.dev>