Introduce the concept of age on codeable object
This is a low effort, mostly correct, way to ensure we encode and keep
the most recent object. It's especially important when an object is
pointed from multiple location but only mutate one of the location. For
example, mutating the PBXProject without changing the container portal
of the PBXContainerItemProxy where most object would point back to the
PBXProject. Reencoding the objects may use the older version
dependending on the traversing. Instead, we track the "age" of the
object and deduplicate by keeping the newers object.

Signed-off-by: Daniel Lacasse <daniel@nokee.dev>