Make a small distinction between registered/created identity
Element registered via ModelMap can be "realized now" as they will be in
a "registered state" so we attach the underlying provider and decide on
the realization then. On the other side, element registered via Gradle
containers will be invisible to us (the ModelMap) until they are created
(we are talking strictly about elements that have a factory registered
via ModelMap as we need the glue the factory introduce). The element
will only become visible once it's realized via the factory glue at
which point the element is already in the process of being realized so
we don't care about attaching the element provider anymore as it only
serves the purpose of realizing the underlying object.

Signed-off-by: Daniel Lacasse <daniel@nokee.dev>